move thee user file into the authentication folder
