+++ server/App.js
@@ -6,7 +6,7 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
+var usersRouter = require('./authentication/users');
 var loginRouter = require('./authentication/authRouter')
 var mailRouter = require('./routes/mail')
